requirements.txt
tweepy==3.8.0
python-dotenv==0.11.0
termcolor==1.1.0
